pyproject.toml
[tool.poetry]
name = "explainable-medical-coding"
version = "0.1.0"
description = ""
authors = ["Joakim <je@corti.ai>"]
license = "The MIT License (MIT)"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
numpy = "^1.26.2"
rich = "^13.7.0"
hydra-core = "^1.3.2"
polars = "^0.20.1"
wandb = "^0.16.1"
scikit-learn = "^1.3.2"
ruff = "^0.1.8"
captum = "^0.7.0"
setuptools = "^69.0.2"
python-dotenv = "^1.0.0"
pre-commit = "^3.6.0"
pydantic = "^2.5.3"
torch = "2.1.1"
transformers = "^4.38.1"
datasets = "^2.17.1"
textattack = "^0.3.9"
tensorflow-hub = "^0.16.1"
tensorflow = "^2.15.0.post1"
ipython = "^8.22.2"
ipykernel = "^6.29.3"
ipywidgets = "^8.1.2"
seaborn = "^0.13.2"
gdown = "^5.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
